Fix variable and color tests, add new test font
Broke the tests with the previous two commits:

Color test broke as SourceCodePro-Regular.ttf contains a few
SVG color glyphs. This is super weird! We didn't catch this
earlier as SVG color glyph recognition was broken, so it
wasn't recognised as a color font. I've now added a miniscule
font with just the letter "A" as TTF test font.

Variable test broke as I forgot to change them to the new
"variable" name :-)