Add lookup type 3, alternate substitution
This needs a change in the way the engine reports on the characters
that belong to a feature. It used to be:

"feat": ["c", "h", "a", "r", "s"]

But is now:

"feat": {
  type: 1,
  input: ["c", "h", "a", "r", "s"]
}

That way the front-end can decide how to show the input chars. E.g.
for lookup 3, we pass the "alternateCount" value, which instructs
the front-end how many alternates of the input character to show.