Create some sort of "flattened type 6" report
In an attempt to make the many possible combinations for a type 6
lookup slightly more comprehensible, we create a summary. This might
result in "false positives", e.g. certain combinations that will
not work together. For example:

Lookup 1:
	Backtrack: a, b, c
	Input: p

Lookup 2:
	Input: q
	Lookahead: x, y, z

Will result in the following summary:

Backtrack: a, b, c
Input: p, q
Lookahead: x, y, z

This would suggest that "aq" or "cpz" would point to a specific
substitution, which isn't so.

So while this is imperfect, it will at least show all characters
involved in this lookup.

As always, Wakamai Fondue reports on actual unicode chars, not
on glyphs that have been replaced in a previous lookup/feature.