Memoize the result of char-based calculations
Stuff was getting intense for fonts with huge charset support. And
we'd call this a few times for any old font. So we cache the result.

Cache busting is, for now, the responsibility of the caller (which in
our case is the site, which resets the fondue when a new font is loaded)