Fix type 6 lookup
Well, this was a doozy! Glad to finally figure this one out. We weren't
combining the involved characters in a meaningul way (and frankly in a
wrong way).

This is a lot better now, and it parses our test font better, so I
updated the test that has been wrong since the last few commits.