Also prevent overwrites for type 3
But I'm now wondering, can we be sure the `alternateCount` is correct?
Or rather, that the index of both arrays are guaranteed to belong to
each other?

I think I need to refactor `alternateCount` to explicitly link the
alternate count to the coverage.

Need to think about this, but merging this for now.