Update type 4
I'm not sure if this would actually prevent overwrites, because the
input is ligature sequences. So could `ffi` be in there twice, for
example? I don't think so, because how would the second entry be
processed if the first is already processed?

But to keep the code in sync with the other lookups, I'm using the
merge approach here too. Manual and automated tests don't seem to mind!