Break up CSS
We now add an options object with everything set
to be included in the final CSS. Just the three
obvious blocks now: fontface, features, and var
instances.

We can add more options later this way, e.g.
to do modern `font-variant` etc.

Next up: computering the site so it works with
the new options thing.